+++ src/charade/charade.navigation.tsx
@@ -1,0 +1,16 @@
+import * as React from 'react'
+import { createNativeStackNavigator } from '@react-navigation/native-stack'
+import { SCREEN } from '../navigation/navigation.constants'
+import { Charade } from './charade'
+import { CharadePlay } from './components/charade-play/charade-play'
+
+const Stack = createNativeStackNavigator()
+
+export const CharadeNavigation = () => {
+  return (
+    <Stack.Navigator screenOptions={{ headerShown: false }}>
+      <Stack.Screen name={SCREEN.CHARADE_MAIN} component={Charade} />
+      <Stack.Screen name={SCREEN.CHARADE_PLAY} component={CharadePlay} />
+    </Stack.Navigator>
+  )
+}