+++ src/charade/components/index.ts
@@ -1,3 +1,2 @@
 export { CharadePlayground } from './charade-playground/charade-playground'
 export { CharadeField } from './charade-field/charade-field'
-export { CharadeHeader } from './charade-header/charade-header'