+++ src/core/header/header.models.ts
@@ -1,0 +1,8 @@
+export type HeaderType = 'more' | 'charade' | 'dictionary' | 'dashboard'
+
+export interface HeaderSideContentConfig {
+  onPress: () => void;
+  onLongPress: () => void;
+  icon: string;
+  indicator: boolean;
+}