+++ src/more/components/listed-option/listed-option.styled.ts
@@ -3,6 +3,7 @@ import styled, { ThemeProps } from 'styled-components/native'
 import { ThemeModel } from '@core/styled/models'
 import { TEXT_SIZE } from '@core/text/text.constants'
 import { getThemeProp } from '@core/styled/theme'
+import { COLOR } from '@core/colors/colors.constants'
 
 export const ListedOptionContainer = styled.View<ThemeProps<ThemeModel>>`
   flex-direction: row;
@@ -15,12 +16,19 @@ export const ListedOptionContainer = styled.View<ThemeProps<ThemeModel>>`
 
 interface ListedOptionHeadlineProps {
   titleSize?: TEXT_SIZE;
+  titleColor?: COLOR;
 }
 
 const getListedOptionHeadline = R.propOr(TEXT_SIZE.M, 'titleSize')
 
+const getListedOptionHeadlineColor = R.ifElse(
+  R.propSatisfies(Boolean, 'titleColor'),
+  R.prop('titleColor'),
+  getThemeProp('textPrimary'),
+)
+
 export const ListedOptionHeadline = styled.Text<ListedOptionHeadlineProps & ThemeProps<ThemeModel>>`
   font-size: ${getListedOptionHeadline}px;
-  color: ${getThemeProp('textPrimary')};
+  color: ${getListedOptionHeadlineColor};
   font-weight: 500;
 `