+++ src/more/hooks/use-more-options.hook.ts
@@ -33,6 +33,8 @@ type Options = [
   MoreOption<undefined>,
   MoreOption<undefined>,
   MoreOption<undefined>,
+  MoreOption<undefined>,
+  MoreOption<undefined>,
 ] | []
 
 interface UseMoreOptions {
@@ -137,6 +139,16 @@ export const useMoreOptions = (
         onChange: () => navigation.navigate(SCREEN.MORE_PLAYGROUND),
         icon: 'checkerboard',
       },
+      {
+        title: localize().help,
+        onChange: () => navigation.navigate(SCREEN.MORE_HELP),
+        icon: 'help',
+      },
+      {
+        title: localize().about_author,
+        onChange: () => navigation.navigate(SCREEN.MORE_AUTHOR),
+        icon: 'account-question',
+      }
     ], [ hapticFeedbackEnabled, nativeSearchEngineEnabled, premium, isPending, languageCode, darkThemeEnabled ])
 
     return { handleDeactivatePremium, getOptions }