+++ src/more/more.navigation.tsx
@@ -4,6 +4,8 @@ import { More } from './more'
 import { SCREEN } from '../navigation/navigation.constants'
 import { Mania } from '../mania/mania'
 import { Playground } from '../playground/playground'
+import { HelpNavigation } from '../help/help.navigation'
+import { Author } from './components/author/author'
 
 const Stack = createNativeStackNavigator()
 
@@ -13,6 +15,8 @@ export const MoreNavigation = () => {
       <Stack.Screen name={SCREEN.MORE_MAIN} component={More} />
       <Stack.Screen name={SCREEN.MORE_MANIA} component={Mania} />
       <Stack.Screen name={SCREEN.MORE_PLAYGROUND} component={Playground} />
+      <Stack.Screen name={SCREEN.MORE_HELP} component={HelpNavigation} />
+      <Stack.Screen name={SCREEN.MORE_AUTHOR} component={Author} />
     </Stack.Navigator>
   )
 }