+++ src/navigation/navigation.helpers.ts
@@ -1,0 +1,7 @@
+export const getNavigationParam = <R>(
+  param: string,
+  navigation: any,
+): R => {
+  const navigationState = navigation?.getState?.()
+  return navigationState?.routes?.[navigationState.index]?.params?.[param]
+}